Refactor HomeAssistantDiscovery service and improve test coverage
